Move View/Taker row helper refactor into separate PR
